Removed name property for Days
The name property was originally used to determine the days and weeks
to skip while downloading new reports. Since it was simply the name
in the popup list on the itts page, this was potentially unreliable,
so the names of reports that should be skipped are now determined
dynamically from the date.