Update PaymentsViewController to use the CurrencyManager to format amounts.
Payment amounts were appearing weirdly in the Payments view, for
example 134.45£. With this update, it will appear as £134.45.
Added two new methods to CurrencyManager to give output in any
specified currency. Edited exiting methods to use those methods so as
not to repeat code.
Removed the separate number formatter in PaymentsViewController.